# business/openapi.yaml
openapi: 3.1.0
info:
  title: rust-invoice-system — business API
  version: "1.0.0"
  description: |
    Invoicing, charging and event delivery.

    ## Two audiences, one port

    Which endpoints need a token is decided by **who is asking**, not by whether the request
    writes. `GET /invoices` is gated and `POST /invoices/{id}/pay` is not, because raising or
    voiding an invoice is the business's own affair, while whoever is being billed has no token
    and cannot be given one.

    - **Operator** — the business's back office. Requires `X-API-Token`.
    - **Payer** — the person being billed. `POST /invoices/{id}/pay` and
      `GET /payment_intents/{id}`, both open by necessity. Neither exposes anything about an
      invoice beyond the charge the caller is already making.
    - **Docker** — `GET /health`, open so the container healthcheck can reach it.

    An unauthenticated request to an operator endpoint gets `401 unauthorized`; a *wrong* token
    gets exactly the same, since telling that caller their header was well formed and merely
    incorrect hands them the one bit they were missing. A path matching no route stays `404`
    rather than `401`, which would reveal which paths exist.

    ## Error format

    Every failure answers with the same shape, so a caller parses failures one way:

    ```json
    { "error": "invoice_not_payable" }
    ```

    One addition, on `POST /invoices/{id}/pay`: once a charge has reached the payment service, every
...

...

...

...
servers:
...
paths:
  /customers:
    post:
      tags: [Customers]
      summary: Create a customer
      operationId: createCustomer
      security: [{ operatorToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCustomer" }
            example: { name: "Acme Corp", email: "billing@acme.test" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "400":
          description: "`invalid_customer` — name or email is blank."
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: "`email_already_exists` — email is unique across customers."
          $ref: "#/components/responses/Error"
        "500": { $ref: "#/components/responses/Internal" }
    get:
      tags: [Customers]
      summary: List every customer
      operationId: listCustomers
      security: [{ operatorToken: [] }]
      responses:
        "200":
          description: All customers, ordered by id (so, by creation).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
...
  /invoices:
    post:
      tags: [Invoices]
      summary: Raise an invoice
      operationId: createInvoice
      description: |
        The caller sends line items and the server totals them; there is no way to supply a total
        directly, so the number stored can never disagree with the lines it came from. The line
        items themselves are **not persisted** — they are input to the sum and then discarded.

        A new invoice is always `draft`. Emits `invoice.created`.
      security: [{ operatorToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoice" }
            example:
              customer_id: "01a04ca6-d4e4-7e81-b6dc-41d699f8fa25"
              due_date: "2026-10-01"
              line_items:
                - { description: "Consulting", quantity: 2, unit_amount_cents: 50000 }
                - { description: "Support", quantity: 1, unit_amount_cents: 2500 }
      responses:
        "201":
          description: Created, in state `draft`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
        "400":
          description: |
            `no_line_items` — the array is empty.
            `invalid_line_item` — blank description, `quantity <= 0`, or a negative unit amount.
            `total_too_large` — the sum overflows a 64-bit integer.
            `invalid_due_date` — not `YYYY-MM-DD`, or not a real calendar day.
            `invalid_customer_id` — not a uuid.
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: "`customer_not_found` — the body is well formed, the customer is not there."
          $ref: "#/components/responses/Error"
        "500": { $ref: "#/components/responses/Internal" }
    get:
      tags: [Invoices]
      summary: List invoices
      operationId: listInvoices
      security: [{ operatorToken: [] }]
      parameters:
        - name: state
          in: query
          required: false
          description: Return only invoices in this state. Omit for all of them.
          schema: { $ref: "#/components/schemas/InvoiceState" }
      responses:
        "200":
          description: Matching invoices, ordered by id.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
        "400":
          description: "`invalid_invoice_state` — not one of the five states."
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
  /invoices/{id}:
    get:
      tags: [Invoices]
      summary: Read one invoice
      operationId: getInvoice
      security: [{ operatorToken: [] }]
      parameters: [{ $ref: "#/components/parameters/InvoiceId" }]
      responses:
        "200":
          description: The invoice.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
        "400":
          description: "`invalid_invoice_id`"
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: "`invoice_not_found`"
          $ref: "#/components/responses/Error"
        "500": { $ref: "#/components/responses/Internal" }
...
  /events:
    get:
      tags: [Webhooks]
      summary: The event log — and how a business reconciles what it missed
      operationId: listEvents
      description: |
        Each entry is **byte for byte the body that was, or would have been, delivered**, so
        catching up here and receiving a webhook are the same information arriving two ways.

        Poll with the id of the last event already processed. `id` is a uuidv7 and its leading
        bits are a timestamp, so ordering by primary key *is* ordering by creation time and
        `after` is a cursor rather than an offset — a page cannot shift under a caller midway
        through reading it.

        A pull, deliberately, rather than a redelivery endpoint: it also covers events raised
        before any endpoint was configured, which have no delivery rows to replay.
      security: [{ operatorToken: [] }]
      parameters:
        - name: after
          in: query
          required: false
          description: Return only events newer than this id.
          schema: { type: string, format: uuid }
        - name: type
          in: query
          required: false
          schema: { $ref: "#/components/schemas/EventType" }
        - name: limit
          in: query
          required: false
          description: 1–1000. Asking for more is refused rather than quietly capped.
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        "200":
          description: Matching events, oldest first. `[]` when nothing is newer — not a 404.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }
        "400":
          description: "`invalid_after` — not a uuid. `invalid_limit` — outside 1–1000."
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    operatorToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: |
        One shared static token, compared in constant time. A missing token and a wrong one get
        the same `401`.
  parameters:
...
    InvoiceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: A failure, in the standard shape.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: "`unauthorized` — no token, or the wrong one."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: unauthorized }
    Internal:
      description: "`internal_error`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: internal_error }
...
  schemas:
    Error:
      type: object
      description: The shape every failure answers with, bar the two documented exceptions.
      properties:
        error:
          type: string
          description: A stable machine-readable code. Never prose.
          example: invoice_not_payable
      required: [error]
    CreateCustomer:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email:
          type: string
          minLength: 1
          description: Unique across customers. Trimmed, not otherwise validated.
      required: [name, email]
    Customer:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
      required: [id, name, email]
    InvoiceState:
      type: string
      enum: [draft, ready, processing, processed, void]
      description: |
        `draft` — being prepared, editable.
        `ready` — finalized and payable.
        `processing` — a charge is in flight; no operator endpoint can move it.
        `processed` — paid. Terminal.
        `void` — written off. Terminal.
    LineItem:
      type: object
      description: Input to the total, then discarded. Not persisted and never returned.
      properties:
        description: { type: string, minLength: 1 }
        quantity:
          type: integer
          minimum: 1
          description: Must be positive — a zero-quantity line says nothing and is likelier a bug.
        unit_amount_cents:
          type: integer
          minimum: 0
          description: Zero is allowed; a comped line still belongs on the invoice.
      required: [description, quantity, unit_amount_cents]
    CreateInvoice:
      type: object
      properties:
        customer_id: { type: string, format: uuid }
        due_date:
          type: string
          format: date
          description: Exactly `YYYY-MM-DD`. Other formats are refused rather than guessed at.
        line_items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/LineItem" }
      required: [customer_id, due_date, line_items]
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        customer_id: { type: string, format: uuid }
        total_cents:
          type: integer
          minimum: 0
          description: Minor units, summed from the line items by the server.
        due_date: { type: string, format: date }
        state: { $ref: "#/components/schemas/InvoiceState" }
      required: [id, customer_id, total_cents, due_date, state]
...
    EventType:
      type: string
      enum: [invoice.created, invoice.paid, invoice.payment_failed]
      description: A receiver must ignore types it does not recognise; the vocabulary only grows.
    Event:
      type: object
      description: The delivered envelope, byte for byte.
      properties:
        id:
          type: string
          format: uuid
          description: Also the `webhook-id` header, and the receiver's deduplication key.
        type: { $ref: "#/components/schemas/EventType" }
        created_at: { type: string, format: date-time }
        data:
          type: object
          properties:
            invoice: { $ref: "#/components/schemas/Invoice" }
            payment_intent_id:
              type: string
              description: Present on `invoice.paid` and `invoice.payment_failed`.
          required: [invoice]
      required: [id, type, created_at, data]
...